Add publish selector test and remove debug output
