+++ api/FlowExt.api
@@ -210,9 +210,7 @@ public abstract interface annotation class com/hoc081098/flowext/PublishSelector
 }
 
 public final class com/hoc081098/flowext/PublishWithSelectorKt {
-	public static final fun main (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public static synthetic fun main ([Ljava/lang/String;)V
-	public static final fun publish (Lkotlinx/coroutines/flow/Flow;Lkotlin/jvm/functions/Function2;)Lkotlinx/coroutines/flow/Flow;
+        public static final fun publish (Lkotlinx/coroutines/flow/Flow;Lkotlin/jvm/functions/Function2;)Lkotlinx/coroutines/flow/Flow;
 }
 
 public final class com/hoc081098/flowext/RaceKt {
